move test to InferringGravsearchTypeInspectorE2ESpec
